Fix Hashing layer compatibility when dtype=float and output_mode=int
When saving a Hashing layer in a model previously, it would save with
a floating point dtype, even if the layer was input strings and outputing
ints. We need to handle this case and clear the setting to avoid an
error on saved configs when output_mode="int". We will ignore floating
point dtype in that case.

PiperOrigin-RevId: 402354163